Tests refactored, replaced mocked factories with factory_boy's real ones.
