+++ .gitignore
@@ -55,5 +55,5 @@ docs/_build/
 target/
 
 .idea/
-
+.venv/
 test.sh
