+++ tests/mocking/mocks.py
@@ -1,0 +1,11 @@
+class MockObject(object):
+    pass
+
+
+class MockFactory(object):
+    @classmethod
+    def create(cls):
+        return MockObject()
+
+
+
