+++ tests/urls.py
@@ -1,32 +1,27 @@
 from rest_framework import routers
 
-from tests import mocks
-
+from tests.app.views import StuffViewSet, StuffHyperlinkedViewSet, RelatedStuffViewSet, RelatedStuffHyperlinkedViewSet
 
 router = routers.DefaultRouter()
 
-router.register(r'stuff',
-                mocks.StuffViewSet,
+router.register(r'stuff', StuffViewSet,
                 base_name='stuff')
 
-router.register(r'stuff-linked',
-                mocks.StuffHyperlinkedViewSet,
+router.register(r'stuff-linked', StuffHyperlinkedViewSet,
                 base_name='stuff-linked')
 
-router.register(r'related-stuff',
-                mocks.RelatedStuffViewSet,
+router.register(r'related-stuff', RelatedStuffViewSet,
                 base_name='relatedstuff')
 
-router.register(r'related-stuff-linked',
-                mocks.RelatedStuffHyperlinkedViewSet,
+router.register(r'related-stuff-linked', RelatedStuffHyperlinkedViewSet,
                 base_name='relatedstuff-linked')
 
-router.register(r'many-related-stuff',
-                mocks.RelatedStuffViewSet,
+router.register(r'many-related-stuff', RelatedStuffViewSet,
                 base_name='manyrelatedstuff')
 
-router.register(r'many-related-stuff-linked',
-                mocks.RelatedStuffHyperlinkedViewSet,
+router.register(r'many-related-stuff-linked', RelatedStuffHyperlinkedViewSet,
                 base_name='manyrelatedstuff-linked')
 
 urlpatterns = router.urls
+print(urlpatterns)
+print('************************************************************************************************************************************************************************************************************************************')
